requirements.txt
kafka-python==2.0.1
environs==7.3.1
ujson==2.0.3
vk_api==11.8.0
black==19.10b0
